# business/src/main/java/truespot/business/dto/AreaDTO.java
package truespot.business.dto;

import lombok.Data;

@Data
public class AreaDTO {

    private long id;
    private String name;
    private String description;
    private String orientation;
    private int roadNumber;
    private int height;

}