# front/truespotFront/src/model/property.ts
/**
 * Swagger Maven Plugin Sample
 * This is a sample.
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Xml } from './xml';


export interface Property {
    position?: number;
    example?: any;
    vendorExtensions?: { [key: string]: any; };
    title?: string;
    xml?: Xml;
    readOnly?: boolean;
    allowEmptyValue?: boolean;
    type?: string;
    format?: string;
    description?: string;
}