# model/src/main/resources/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "This is a sample.",
    "version" : "v1",
    "title" : "Swagger Maven Plugin Sample",
    "termsOfService" : "http://www.github.com/kongchen/swagger-maven-plugin",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "tags" : [ {
    "name" : "data"
  } ],
  "paths" : {
    "/api/area" : {
      "get" : {
        "tags" : [ "data" ],
        "summary" : "",
        "description" : "",
        "operationId" : "area",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Area"
              }
            }
          },
          "400" : {
            "description" : "Bad Request ",
            "schema" : {
              "$ref" : "#/definitions/Area"
            }
          }
        }
      }
    },
    "/api/department" : {
      "get" : {
        "tags" : [ "data" ],
        "summary" : "",
        "description" : "",
        "operationId" : "department",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Department"
              }
            }
          },
          "400" : {
            "description" : "Bad Request ",
            "schema" : {
              "$ref" : "#/definitions/Department"
            }
          }
        }
      }
    },
    "/api/road" : {
      "get" : {
        "tags" : [ "data" ],
        "summary" : "",
        "description" : "",
        "operationId" : "road",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ClimbingRoad"
              }
            }
          },
          "400" : {
            "description" : "Bad Request ",
            "schema" : {
              "$ref" : "#/definitions/ClimbingRoad"
            }
          }
        }
      }
    },
    "/api/roadType" : {
      "get" : {
        "tags" : [ "data" ],
        "summary" : "",
        "description" : "",
        "operationId" : "roadType",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ClimbingRoadType"
              }
            }
          },
          "400" : {
            "description" : "Bad Request ",
            "schema" : {
              "$ref" : "#/definitions/ClimbingRoadType"
            }
          }
        }
      }
    },
    "/api/spot" : {
      "get" : {
        "tags" : [ "data" ],
        "summary" : "",
        "description" : "",
        "operationId" : "spot",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Spot"
              }
            }
          },
          "400" : {
            "description" : "Bad Request ",
            "schema" : {
              "$ref" : "#/definitions/Spot"
            }
          }
        }
      }
    },
    "/api/topo" : {
      "get" : {
        "tags" : [ "data" ],
        "summary" : "",
        "description" : "",
        "operationId" : "topo",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Topo"
              }
            }
          },
          "400" : {
            "description" : "Bad Request ",
            "schema" : {
              "$ref" : "#/definitions/Topo"
            }
          }
        }
      }
    },
    "/api/user" : {
      "get" : {
        "tags" : [ "data" ],
        "summary" : "",
        "description" : "",
        "operationId" : "user",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/User"
              }
            }
          },
          "400" : {
            "description" : "Bad Request ",
            "schema" : {
              "$ref" : "#/definitions/User"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "Area" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "orientation" : {
          "type" : "string"
        },
        "roadNumber" : {
          "type" : "integer",
          "format" : "int32"
        },
        "height" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "ClimbingRoad" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "area" : {
          "$ref" : "#/definitions/Area"
        },
        "name" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "difficulty" : {
          "$ref" : "#/definitions/Difficulty"
        },
        "type" : {
          "$ref" : "#/definitions/ClimbingRoadType"
        }
      }
    },
    "ClimbingRoadType" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string"
        }
      }
    },
    "Department" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "number" : {
          "type" : "integer",
          "format" : "int32"
        },
        "name" : {
          "type" : "string"
        }
      }
    },
    "Difficulty" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "number" : {
          "type" : "integer",
          "format" : "int32"
        },
        "letter" : {
          "type" : "string"
        },
        "symbol" : {
          "type" : "string"
        }
      }
    },
    "Property" : {
      "type" : "object",
      "properties" : {
        "position" : {
          "type" : "integer",
          "format" : "int32"
        },
        "example" : {
          "type" : "object"
        },
        "vendorExtensions" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "object"
          }
        },
        "title" : {
          "type" : "string"
        },
        "xml" : {
          "$ref" : "#/definitions/Xml"
        },
        "readOnly" : {
          "type" : "boolean"
        },
        "allowEmptyValue" : {
          "type" : "boolean"
        },
        "type" : {
          "type" : "string"
        },
        "format" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        }
      }
    },
    "Response" : {
      "type" : "object",
      "properties" : {
        "description" : {
          "type" : "string"
        },
        "schema" : {
          "$ref" : "#/definitions/Property"
        },
        "examples" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "object"
          }
        },
        "headers" : {
          "type" : "object",
          "additionalProperties" : {
            "$ref" : "#/definitions/Property"
          }
        }
      }
    },
    "Role" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string"
        }
      }
    },
    "Spot" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "topo" : {
          "$ref" : "#/definitions/Topo"
        },
        "name" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "nearestCity" : {
          "type" : "string"
        },
        "carAccess" : {
          "type" : "boolean"
        },
        "carParking" : {
          "type" : "boolean"
        },
        "accessDescription" : {
          "type" : "string"
        },
        "nearestHospital" : {
          "type" : "string"
        },
        "supplyComment" : {
          "type" : "string"
        }
      }
    },
    "Topo" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string"
        },
        "user" : {
          "$ref" : "#/definitions/User"
        },
        "department" : {
          "$ref" : "#/definitions/Department"
        }
      }
    },
    "User" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "roles" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Role"
          }
        },
        "topo" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Topo"
          }
        },
        "name" : {
          "type" : "string"
        },
        "realName" : {
          "type" : "string"
        },
        "age" : {
          "type" : "integer",
          "format" : "int32"
        },
        "gender" : {
          "type" : "string"
        },
        "pseudo" : {
          "type" : "string"
        },
        "password" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "phoneNumber" : {
          "type" : "string"
        },
        "admin" : {
          "type" : "boolean"
        }
      }
    },
    "Xml" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "namespace" : {
          "type" : "string"
        },
        "prefix" : {
          "type" : "string"
        },
        "attribute" : {
          "type" : "boolean"
        },
        "wrapped" : {
          "type" : "boolean"
        }
      }
    }
  }
}